refactor(bench): merge micro-benchmarks into one suite
Consolidate the benchmark components from 7 benchmarks plus 5 internal
libraries down to 3 benchmarks plus 1 internal library.

- Merge copy-at-bench, growable-bench, unboxed-bench, multi-store-scan-bench
  and worklist-resume-bench into pure-borrow-bench, a single -N1 -T executable
  driven by tasty-discover from bench/suite. Each sub-suite is a module under
  PureBorrow.Bench exporting a test_... :: [Benchmark]. The driver injects
  tasty-bench's reporters via --ingredient, kept composed in
  PureBorrow.Bench.Ingredients so --csv and --svg stay reachable, with
  listingTests ordered first so --list-tests keeps working.
- Drop the four single-module internal libraries that existed only to feed one
  benchmark each.
- Remove the artifact-runner executable, the artifact flag, the artifact
  Dockerfile and the Artifact Evaluation workflow; they were historical PLDI
  artifact-evaluation tooling. With artifact-runner gone, QSort and Fft have no
  consumer outside their own benchmark, so they are inlined into bench/qsort.hs
  and bench/fft.hs as self-contained main modules. Benchmark names there are
  unchanged.
- Rename the surviving internal library to test-bench-common: it now holds only
  the kernels that pure-borrow-test and pure-borrow-inspection also exercise,
  so a single shared library keeps the inspected Core and the timed object code
  the same artifact.
