+++ AGENTS.md
@@ -31,8 +31,8 @@ read it before designing changes to the core.
 
 ## Build / test / bench / run
 
-`cabal.project` sets the `+examples` flag and `cabal.project.local` adds `+artifact` and
-enables tests+benchmarks, so locally every component below is buildable.
+`cabal.project` sets the `+examples` flag and `cabal.project.local` enables
+tests+benchmarks, so locally every component below is buildable.
 
 ```bash
 cabal build all                       # build the library + all enabled components
@@ -44,6 +44,7 @@ cabal test pure-borrow-doctests       # doctests (needs GHC >= 9.12.3)
 
 cabal bench qsort-bench               # parallel quicksort benchmark (tasty-bench)
 cabal bench fft-bench                 # parallel FFT benchmark (tasty-bench)
+cabal bench pure-borrow-bench         # all single-threaded micro-benchmarks (needs -j1)
 
 cabal run qsort -- --help             # quicksort demo executable (needs +examples)
 cabal run fft   -- --help             # FFT demo executable (needs +examples)
@@ -79,15 +80,34 @@ the Cabal-selected compiler. Keep errors that GHC does defer in `TypingCases`.
 
 ### Benchmarks & profiling
 
-Benchmarks are `tasty-bench` executables; pass options through cabal, e.g.
+There are exactly three benchmark suites, all `tasty-bench` executables:
+
+- **`qsort-bench`** (`bench/qsort.hs`) and **`fft-bench`** (`bench/fft.hs`) — the two
+  parallel whole-algorithm suites from the paper. Each is a single self-contained
+  `main` module with no `other-modules`; they run under `-N` and take a tasty `--size`
+  option, so they stay separate.
+- **`pure-borrow-bench`** (`bench/suite/`) — every single-threaded micro-benchmark
+  (`copy-at`, `growable`, `unboxed`, `multi-store-scan`, `worklist-*`), merged into one
+  `-N1 -T` executable. Select a subset with tasty's `-p`; `-j1` is mandatory.
+
+`bench/suite/Main.hs` is only the `tasty-discover` driver; each sub-suite is a module under
+`bench/suite/PureBorrow/Bench/` exporting a single `test_… :: [Benchmark]`. Kernels that
+`pure-borrow-test` or `pure-borrow-inspection` also exercise stay in the
+`test-bench-common` internal library and are re-exported by a thin module here; kernels
+used only by the benchmark live under `bench/suite/` directly. Do **not** add a
+per-benchmark internal library — that pattern was removed deliberately, and a library is
+warranted only when a kernel needs a consumer outside its own benchmark.
+
+The driver's `--ingredient` flags rebuild `Test.Tasty.Bench.benchIngredients` (which
+`tasty-discover` cannot use directly) out of `listingTests` and
+`PureBorrow.Bench.Ingredients.benchReporter`; the reporters must stay composed or
+`--csv`/`--svg` become unreachable. `tasty-discover` imports only modules that export a
+discovered binding, so `Ingredients` needs no `--ignores`.
+
+Pass options through cabal, e.g.
 `cabal bench qsort-bench --benchmark-options='--csv bench-results/qsort.csv -j1 --time-mode=wall +RTS -N -s'`.
 For profiled runs use the dedicated project file: `cabal --project-file=cabal-bench.project ...`
 (enables `-fprof-late`/`-fprof-auto` profiling). CSV/plots land in `bench-results/`.
-
-### Artifact runner
-
-With `+artifact` (on locally), `cabal run artifact-runner -- <bench|demo|quick> ...` drives
-the paper's benchmarks/demos and produces CSVs.
 
 ## Architecture
 
@@ -211,8 +231,8 @@ parallel `qsort` (budgeted `parBO`; the heavier version is `qsortDC` above).
 - **Strict warnings** (`common defaults` in `pure-borrow.cabal`): `-Wall -Wcompat
   -Wunused-packages …`. `-Wunused-packages` means a now-unused `build-depends` entry breaks
   the build — drop deps you stop using.
-- **Flags:** `examples` (demo executables + `demo-impl` lib) and `artifact` (artifact-runner)
-  are `manual`, default `False`; enabled locally via `cabal.project`/`cabal.project.local`.
+- **Flags:** `examples` (demo executables + `demo-impl` lib) is the only flag; it is
+  `manual`, default `False`, and enabled locally via `cabal.project`.
 - CI (`.github/workflows/haskell.yml`) runs a GHC matrix (9.10.3/9.12.4/9.14.1 via
   `ci/configs/*.project`): build, all test suites, `cabal check`, Haddock-for-Hackage, and a
   fourmolu check.
