--- bench/copy-at.hs
@@ -1,6 +1,0 @@
-module Main (main) where
-
-import PureBorrow.Internal.Bench.CopyAt (defaultMain)
-
-main :: IO ()
-main = defaultMain
