--- bench/growable.hs
@@ -1,6 +1,0 @@
-module Main (main) where
-
-import PureBorrow.Internal.Bench.Growable (defaultMain)
-
-main :: IO ()
-main = defaultMain
