--- bench/multi-store-scan.hs
@@ -1,6 +1,0 @@
-module Main (main) where
-
-import PureBorrow.Internal.Bench.MultiStoreScan qualified as MultiStoreScan
-
-main :: IO ()
-main = MultiStoreScan.defaultMain
