--- bench/unboxed.hs
@@ -1,4 +1,0 @@
-import PureBorrow.Internal.Bench.Unboxed (defaultMain)
-
-main :: IO ()
-main = defaultMain
