--- bench/worklist-resume.hs
@@ -1,6 +1,0 @@
-module Main (main) where
-
-import PureBorrow.Internal.Bench.Worklist.Resume qualified as Worklist
-
-main :: IO ()
-main = Worklist.defaultMain
